+++ alembic/versions/4c7e55213558_remove_invalid_cell_towers.py
@@ -29,6 +29,9 @@ def upgrade():
     stmt = 'SELECT max(time) FROM stat WHERE `key` = 2'
     max_date = bind.execute(stmt).fetchone()[0]
 
+    if max_date is None:
+        return
+
     # update stat set value = value - <deleted_rows>
     # where `key` = 2 and time = <max_date>
     stmt = ('UPDATE stat SET value = value - {deleted} '